Abstracts away dynamic page data, switches to server-side data fetching, adds metadata
